src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==0.12.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
blis==0.7.4
    # via
    #   spacy
    #   thinc
cached-property==1.5.2
    # via h5py
cachetools==4.2.2
    # via google-auth
catalogue==2.0.4
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via typer
cymem==2.0.5
    # via
    #   preshed
    #   spacy
    #   thinc
flatbuffers==1.12
    # via tensorflow
gast==0.4.0
    # via tensorflow
google-auth-oauthlib==0.4.4
    # via tensorboard
google-auth==1.30.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-pasta==0.2.0
    # via tensorflow
gpt-2-simple==0.7.2
    # via -r requirements.in
grpcio==1.34.1
    # via
    #   tensorboard
    #   tensorflow
h5py==3.1.0
    # via tensorflow
idna==2.10
    # via requests
importlib-metadata==4.0.1
    # via markdown
jinja2==3.0.0
    # via spacy
keras-nightly==2.5.0.dev2021032900
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
markdown==3.3.4
    # via tensorboard
markupsafe==2.0.0
    # via jinja2
murmurhash==1.0.5
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==1.19.5
    # via
    #   blis
    #   gpt-2-simple
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   pandas
    #   spacy
    #   tensorboard
    #   tensorflow
    #   thinc
oauthlib==3.1.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==20.9
    # via spacy
pandas==1.2.4
    # via -r requirements.in
pathy==0.5.2
    # via spacy
pillow==8.2.0
    # via -r requirements.in
preshed==3.0.5
    # via
    #   spacy
    #   thinc
protobuf==3.17.0
    # via
    #   tensorboard
    #   tensorflow
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pydantic==1.7.4
    # via
    #   spacy
    #   thinc
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.1
    # via pandas
pytz==2021.1
    # via pandas
regex==2021.4.4
    # via
    #   -r requirements.in
    #   gpt-2-simple
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   gpt-2-simple
    #   requests-oauthlib
    #   smart-open
    #   spacy
    #   tensorboard
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   protobuf
    #   python-dateutil
    #   tensorflow
smart-open==3.0.0
    # via pathy
spacy-legacy==3.0.5
    # via spacy
spacy==3.0.6
    # via -r requirements.in
srsly==2.4.1
    # via
    #   spacy
    #   thinc
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.5.0
    # via tensorflow
tensorflow-estimator==2.5.0
    # via tensorflow
tensorflow==2.5.0
    # via -r requirements.in
termcolor==1.1.0
    # via tensorflow
thinc==8.0.3
    # via spacy
toposort==1.6
    # via gpt-2-simple
tqdm==4.60.0
    # via
    #   gpt-2-simple
    #   spacy
typer==0.3.2
    # via
    #   pathy
    #   spacy
typing-extensions==3.7.4.3
    # via
    #   catalogue
    #   importlib-metadata
    #   spacy
    #   tensorflow
    #   thinc
urllib3==1.26.4
    # via requests
wasabi==0.8.2
    # via
    #   spacy
    #   thinc
werkzeug==2.0.0
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow
zipp==3.4.1
    # via
    #   catalogue
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
